requirements-agent.txt
agent-framework>=1.0.0rc5
azure-identity>=1.25.0
python-dotenv>=1.0
